+++ Sprint_2/doc/Change log.md
@@ -9,5 +9,33 @@ This file will contain all notable changes to this project.
 - เพิ่มรายละเอียด UI การจองและการเลือกจำนวนสัมภาระ
 - เพิ่มโฟลเดอร์ Sprint 2 โดยประกอบไปด้วยโฟลเดอร์ code doc sprint backlog และ test
 ### AI Declare
-- ในการเพิ่มรายละเอียด UI การจองและการเลือกจำนวนสัมภาระ ได้มีการนำ Gemini 3.1 Pro มาช่วยในการเขียนปรับปรุงโค้ด
+- ใช้ ChatGPT ช่วยออกแบบโครงสร้าง media แบบ Json array
+- ใช้ ChatGPT ช่วยออกแบบ validation URL และ logic จำกัดไฟล์
+- ใช้ ChatGPT ช่วย debug flow completedAt และการคำนวณ 7 วัน
+- ใช้ ChatGPT ช่วยจัดโครงสร้าง Multer และ multiple upload
+- ใช้ ChatGPT ช่วยออกแบบระบบ Fullscreen Video ใน Modal
+---
+## extraCharges V.01 - [Nattawadee]
+### Update 1 Mar 2026
+File includes route.controller.js route.service.js route.validation.js \createTrip\index.vue
+- เพิ่มช่อง extraCharges ในหน้าสร้างเส้นทาง และเชื่อมกับ database
+### AI Declare ใช้ chatgpt ช่วยหาจุดผิดพลาด เนื่องจากข้อมูลไม่บันทึกลง database
+---
+## [Vesion 1.0.0] - [Chatsiri]
+### Update 1 March 2026
+- เพิ่ม html "เงื่อนไขเพิ่มเติม route" ใน page\findTrip
+- เพิ่ม html "เงื่อนไขเพิ่มเติม route" ใน page\myRoute
+- เพิ่ม html "เงื่อนไขเพิ่มเติม trip" ใน page\myTrip
+- รอ api backend เพื่อแสดงข้อมูล
+
+### AI Declare
+---
+## frontend & route.service.js - [Ratchapoom]
+### Update 2 March 2026
+- แก้ไข index.vue\findTrip เพื่อให้แสดงเงื่อนไขเพิ่มเติม (routeExtraCharge) ในรายละเอียดเส้นทาง
+- แก้ไข index.vue\myRoute เพื่อให้แสดงเงื่อนไขเพิ่มเติม (routeExtraCharge) ในรายละเอียดเส้นทาง
+- แก้ไข index.vue\myTrip เพื่อให้แสดงเงื่อนไขเพิ่มเติม (routeExtraCharge) ในรายละเอียดเส้นทาง
+- ใน route.service.js เพิ่ม routeExtraCharge แสดงเงื่อนไขเพิ่มเติม (routeExtraCharge)
+### AI Declare
+- ใช้ ChatGPT 4.1 ช่วยหาจุดผิดพลาดเนื่องจากข้อมูลไม่บันทึกลง database
 ---